+++ src/main/java/org/truffleruby/core/module/ModuleNodes.java
@@ -53,6 +53,7 @@ import org.truffleruby.core.cast.ToPathNodeGen;
 import org.truffleruby.core.cast.ToStrNode;
 import org.truffleruby.core.cast.ToStringOrSymbolNodeGen;
 import org.truffleruby.core.constant.WarnAlreadyInitializedNode;
+import org.truffleruby.core.kernel.KernelNodes;
 import org.truffleruby.core.method.MethodFilter;
 import org.truffleruby.core.module.ModuleNodesFactory.ClassExecNodeFactory;
 import org.truffleruby.core.module.ModuleNodesFactory.SetMethodVisibilityNodeGen;
@@ -613,16 +614,13 @@ public abstract class ModuleNodes {
         @TruffleBoundary
         private CodeLoader.DeferredCall classEvalSource(DynamicObject module, DynamicObject rubySource, String file, int line) {
             assert RubyGuards.isRubyString(rubySource);
-            final String code = StringOperations.getString(rubySource);
+            String code = StringOperations.getString(rubySource);
 
             final MaterializedFrame callerFrame = getContext().getCallStack().getCallerFrameIgnoringSend()
                     .getFrame(FrameInstance.FrameAccess.MATERIALIZE).materialize();
-            Encoding encoding = Layouts.STRING.getRope(rubySource).getEncoding();
-
-            // TODO (pitr 15-Oct-2015): fix this ugly hack, required for AS, copy-paste
-            final char[] emptyLines = new char[Math.max(line - 1, 0)];
-            Arrays.fill(emptyLines, '\n');
-            Source source = Source.newBuilder(new String(emptyLines) + code).name(file.intern()).mimeType(RubyLanguage.MIME_TYPE).build();
+            final Encoding encoding = Layouts.STRING.getRope(rubySource).getEncoding();
+            code = KernelNodes.EvalNode.offsetSource("class/module_eval", code, file, line);
+            Source source = Source.newBuilder(code).name(file.intern()).mimeType(RubyLanguage.MIME_TYPE).build();
 
             final RubyRootNode rootNode = getContext().getCodeLoader().parse(source, encoding, ParserContext.MODULE, callerFrame, true, this);
             final DeclarationContext declarationContext = new DeclarationContext(Visibility.PUBLIC, new FixedDefaultDefinee(module));